Pull request #191: TFP-917 cloudlets edgegrid golang support for load balancing activations
Merge in DEVEXP/akamaiopen-edgegrid-golang from feature/TFP-917-cloudlets-edgegrid-golang-support-for-load-balancing-activations to feature/sp-cloudlets

Squashed commit of the following:

commit 6bc402557a8100a57c9555c5ef34b07d5f670885
Author: Piyush Kaushik <piykaush@akamai.com>
Date:   Fri Sep 10 15:16:11 2021 +0200

    TFP-917 Correct version type

commit f18fe115a8324d0717b13cec92cdb807e559b2f3
Author: Piyush Kaushik <piykaush@akamai.com>
Date:   Tue Sep 7 21:46:12 2021 +0200

    TFP-917 Cloudlets Edgegrid golang support for load balancing activations

+++ pkg/cloudlets/loadbalancer_version.go
@@ -80,7 +80,7 @@ type (
 		LastModifiedDate string            `json:"lastModifiedDate,omitempty"`
 		LivenessSettings *LivenessSettings `json:"livenessSettings,omitempty"`
 		OriginID         string            `json:"originID,omitempty"`
-		Version          int               `json:"version,omitempty"`
+		Version          int64             `json:"version,omitempty"`
 		Warnings         []Warning         `json:"warnings,omitempty"`
 	}
 
@@ -101,7 +101,7 @@ type (
 	UpdateLoadBalancerVersionRequest struct {
 		OriginID            string
 		ShouldValidate      bool
-		Version             int
+		Version             int64
 		LoadBalancerVersion LoadBalancerVersion
 	}
 )
